migrate to rails 5 API
