+++ app/controllers/data_centers_controller.rb
@@ -1,4 +1,4 @@
-class Api::DataCentersController < Api::BaseController
+class DataCentersController < ApplicationController
   before_filter :authenticate_user_from_token!, :set_include
 
   def set_include
