+++ app/controllers/milestones_controller.rb
@@ -1,4 +1,4 @@
-class Api::MilestonesController < Api::BaseController
+class MilestonesController < ApplicationController
   def index
     @milestones = Milestone.where(params.merge(github_token: ENV['GITHUB_PERSONAL_ACCESS_TOKEN']))
     render jsonapi: @milestones[:data], meta: @milestones[:meta]
