+++ config/environments/development.rb
@@ -9,44 +9,32 @@ Rails.application.configure do
   # Do not eager load code on boot.
   config.eager_load = false
 
-  # Show full error reports and disable caching.
-  config.consider_all_requests_local       = true
-  config.action_controller.perform_caching = false
+  # Show full error reports.
+  config.consider_all_requests_local = true
+
+  # config.action_controller.perform_caching = true
+  config.action_controller.perform_caching = true
+  config.cache_store = :dalli_store
 
   # Don't care if the mailer can't send.
-  config.action_mailer.raise_delivery_errors = false
+  # config.action_mailer.raise_delivery_errors = false
+  #
+  # config.action_mailer.perform_caching = false
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
-  # Raise an error on page load if there are pending migrations.
-  # config.active_record.migration_error = :page_load
-
-  # Debug mode disables concatenation and preprocessing of assets.
-  # This option may cause significant delays in view rendering with a large
-  # number of complex assets.
-  config.assets.debug = true
-
-  # Asset digests allow you to set far-future HTTP expiration dates on all assets,
-  # yet still be able to expire them through the digest params.
-  config.assets.digest = false
-
-  # Use a different cache store
-  # dalli uses ENV['MEMCACHE_SERVERS']
-  ENV['MEMCACHE_SERVERS'] ||= ENV['HOSTNAME']
-  config.cache_store = :dalli_store, nil, { :namespace => ENV['APPLICATION'], :compress => true }
-
-  # Adds additional error checking when serving assets at runtime.
-  # Checks for improperly declared sprockets dependencies.
-  # Raises helpful error messages.
-  config.assets.raise_runtime_errors = true
-
-  # See everything in the log (default is :info)
-  config.log_level = :debug
-
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
-  # config.web_console.whitelisted_ips = ENV['TRUSTED_IP']
+
+  # Use an evented file watcher to asynchronously detect changes in source code,
+  # routes, locales, etc. This feature depends on the listen gem.
+  config.file_watcher = ActiveSupport::EventedFileUpdateChecker
 end
 
 BetterErrors::Middleware.allow_ip! ENV['TRUSTED_IP']
+
+HttpLog.configure do |config|
+  config.logger = ActiveSupport::Logger.new(STDOUT)
+  config.severity = Logger::Severity::DEBUG
+end
