--- config/initializers/httplog.rb
@@ -1,4 +1,0 @@
-HttpLog.configure do |config|
-  config.logger = Rails.logger
-  config.log_headers = (ENV['LOG_LEVEL'] == "debug")
-end
