--- config/initializers/session_store.rb
@@ -1,2 +1,0 @@
-# Be sure to restart your server when you modify this file.
-Rails.application.config.session_store :cookie_store, key: ENV['SESSION_KEY']
