+++ config/initializers/version.rb
@@ -1,3 +1,3 @@
 module Spinone
-  VERSION = "3.0"
+  VERSION = "2.0"
 end
