--- public/v1/swagger.json
@@ -1,125 +1,0 @@
-{
-    "swagger": "2.0",
-    "info": {
-        "version": "0.1.0",
-        "title": "DataCite Labs API",
-        "description": "The DataCite Labs API is currently not suited for production use. Please contact DataCite support for more information.\n",
-        "contact": {
-            "name": "DataCite Support",
-            "email": "support@datacite.org"
-        }
-    },
-    "paths": {
-        "/works": {
-            "get": {
-                "description": "Gets `Work` objects. Queries the DataCite [MDS](https://mds.datacite.org). The following parameters also work with [DataCite Event Data](https://eventdata.datacite.org):\n\n* id\n* publisher-id\n* source-id (only Event Data)\n* relation-type-id (only Event Data)\n\nDataCite Event Data also contains DOI names from other DOI registration agencies, as well as resources \nwithout a DOI. \n\nThe following optional query param determine the offset, size and sorting of the results returned:\n\n* offset\n* rows\n* sort\n",
-                "parameters": [
-                    {
-                        "name": "q",
-                        "in": "query",
-                        "description": "Query string",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "id",
-                        "in": "query",
-                        "description": "Work ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "member-id",
-                        "in": "query",
-                        "description": "Member ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "publisher-id",
-                        "in": "query",
-                        "description": "Publisher ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "resource-type-id",
-                        "in": "query",
-                        "description": "Resource Type ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "source-id",
-                        "in": "query",
-                        "description": "Source ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "relation-type-id",
-                        "in": "query",
-                        "description": "Relation Type ID",
-                        "required": false,
-                        "type": "string"
-                    },
-                    {
-                        "name": "offset",
-                        "in": "query",
-                        "description": "Result offset, defaults to **0**",
-                        "required": false,
-                        "type": "number"
-                    },
-                    {
-                        "name": "rows",
-                        "in": "query",
-                        "description": "Size of returned array, defaults to **25**",
-                        "required": false,
-                        "type": "number"
-                    },
-                    {
-                        "name": "sort",
-                        "in": "query",
-                        "description": "Result sorting field, defaults to **score** when **q** parameter is used, \n**minted** otherwise\n",
-                        "required": false,
-                        "type": "number"
-                    },
-                    {
-                        "name": "order",
-                        "in": "query",
-                        "description": "Sort order, defaults to **desc**",
-                        "required": false,
-                        "type": "string"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "Successful response",
-                        "schema": {
-                            "title": "ArrayOfWorks",
-                            "type": "array",
-                            "items": {
-                                "title": "Work",
-                                "type": "object",
-                                "properties": {
-                                    "id": {
-                                        "type": "string"
-                                    },
-                                    "title": {
-                                        "type": "string"
-                                    },
-                                    "container-title": {
-                                        "type": "string"
-                                    }
-                                }
-                            }
-                        }
-                    },
-                    "404": {
-                        "description": "Not found"
-                    }
-                }
-            }
-        }
-    }
-}
